+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Fixed:
+  * [GITHUB-3](https://github.com/claranet/terraform-azurerm-regions/issues/3): Fix region Australia Southeast short name, cleanup Australia South
+  * AZ-720: Add `try()` on outputs
+
 # v5.0.0 - 2022-04-05
 
 Added: